access-analyzer/iam-role-findings-resolution/requirements.txt
aws-cdk.core==1.85.0
aws_cdk.aws_s3==1.85.0
aws_cdk.aws_accessanalyzer==1.85.0
aws_cdk.aws_kms==1.85.0
aws_cdk.aws_iam==1.85.0
aws_cdk.aws_lambda==1.85.0
aws_cdk.aws_events==1.85.0
aws_cdk.aws_events_targets==1.85.0
aws_cdk.aws_sqs==1.85.0
aws_cdk.aws_sns==1.85.0
aws_cdk.aws_sns_subscriptions==1.85.0
aws_cdk.aws_stepfunctions==1.85.0
aws_cdk.aws_stepfunctions_tasks==1.85.0
aws_cdk.aws_kms==1.85.0
